+++ CHANGELOG.md
@@ -7,5 +7,7 @@
 ### Fixes:
 - Fixed a bug where party leadership transfers weren't correctly detected due to a typo.
 - Suppressed redundant error popup when joining a party via the fallback method.
+- Made lava hotspots only track smoke particles and water hotspots only dust particles
+- Made lava hotspots require more particles before defining size
 
 ### Back-end: